🚚[#3416] Move "public" appointments stuff to .service
